libphoenix: rename dirs in include/arch and crt0 files
Rename crt0 files to crt0-common.c and crt0.S.
Change dir names in include/arch from arm-imx and armv7 to armv7a and
armv7m respectively, to match names in arch directory.

JIRA: RTOS-712
