+++ arch/armv7a/Makefile
@@ -1,9 +1,9 @@
 #
-# Makefile for libphoenix/arch/arm-imx
+# Makefile for libphoenix/arch/armv7a
 #
 # Copyright 2017, 2020 Phoenix Systems
 # Author: Pawel Pisarczyk
 #
 
 OBJS += $(addprefix $(PREFIX_O)arch/armv7a/, syscalls.o jmp.o signal.o string.o reboot.o)
-CRT0_OBJS += $(addprefix $(PREFIX_O)arch/armv7a/, _start.o)
+CRT0_OBJS += $(addprefix $(PREFIX_O)arch/armv7a/, crt0.o)
