+++ arch/armv7m/Makefile
@@ -1,9 +1,9 @@
 #
-# Makefile for libphoenix/arch/armv7
+# Makefile for libphoenix/arch/armv7m
 #
 # Copyright 2017, 2020 Phoenix Systems
 # Author: Pawel Pisarczyk
 #
 
 OBJS += $(addprefix $(PREFIX_O)arch/armv7m/, syscalls.o jmp.o signal.o string.o reboot.o tls.o)
-CRT0_OBJS += $(addprefix $(PREFIX_O)arch/armv7m/, _start.o)
+CRT0_OBJS += $(addprefix $(PREFIX_O)arch/armv7m/, crt0.o)
