+++ arch/ia32/Makefile
@@ -6,4 +6,4 @@
 #
 
 OBJS += $(addprefix $(PREFIX_O)arch/ia32/, syscalls.o jmp.o signal.o string.o reboot.o)
-CRT0_OBJS += $(addprefix $(PREFIX_O)arch/ia32/, _start.o)
+CRT0_OBJS += $(addprefix $(PREFIX_O)arch/ia32/, crt0.o)
