+++ arch/riscv64/Makefile
@@ -6,4 +6,4 @@
 #
 
 OBJS += $(addprefix $(PREFIX_O)arch/riscv64/, syscalls.o string.o signal.o reboot.o jmp.o)
-CRT0_OBJS += $(addprefix $(PREFIX_O)arch/riscv64/, _start.o)
+CRT0_OBJS += $(addprefix $(PREFIX_O)arch/riscv64/, crt0.o)
