+++ arch/riscv64/crt0.S
