+++ crt0-common.c
