+++ include/arch/armv7a/arch.h
@@ -3,22 +3,22 @@
  *
  * libphoenix
  *
- * Architecture dependent part (arch/armv7)
+ * Architecture dependent part (arch/armv7a)
  *
- * Copyright 2017 Phoenix Systems
- * Author: Pawel Pisarczyk
+ * Copyright 2017, 2018 Phoenix Systems
+ * Author: Pawel Pisarczyk, Aleksander Kaminski
  *
  * This file is part of Phoenix-RTOS.
  *
  * %LICENSE%
  */
 
-#ifndef _LIBPHOENIX_ARCH_ARMV7_ARCH_H_
-#define _LIBPHOENIX_ARCH_ARMV7_ARCH_H_
+#ifndef _LIBPHOENIX_ARCH_ARMV7A_ARCH_H_
+#define _LIBPHOENIX_ARCH_ARMV7A_ARCH_H_
 
-#define __ARCH_STDINT <arch/armv7/stdint.h>
-#define __ARCH_LIMITS <arch/armv7/limits.h>
-#define __ARCH_SYS_TYPES <arch/armv7/types.h>
+#define __ARCH_STDINT    <arch/armv7a/stdint.h>
+#define __ARCH_LIMITS    <arch/armv7a/limits.h>
+#define __ARCH_SYS_TYPES <arch/armv7a/types.h>
 
 #define __MEMCPY
 #define __MEMCMP
@@ -30,6 +30,7 @@
 #define __STRCPY
 #define __STRNCPY
 #define __MEMMOVE
+
 
 #if __ARM_PCS_VFP || (__VFP_FP__ && !__SOFTFP__)
 #if __ARM_FP & 8
@@ -57,14 +58,9 @@ static inline float __ieee754_sqrtf(float x)
 }
 #endif
 
-#define _PAGE_SIZE 0x200
-#define SIZE_PAGE _Pragma("GCC warning \"'SIZE_PAGE' is deprecated. Use _PAGE_SIZE from arch.h or PAGE_SIZE from limits.h (POSIX only)\"") _PAGE_SIZE
+#define _PAGE_SIZE 0x1000
+#define SIZE_PAGE  _Pragma("GCC warning \"'SIZE_PAGE' is deprecated. Use _PAGE_SIZE from arch.h or PAGE_SIZE from limits.h (POSIX only)\"") _PAGE_SIZE
 
-/* FIXME provide libphoenix config to be able to
- * selectively disable/enable features on per
- * project basis.
- * Disabled for now as TLS consumes too much
- * memory to be advantageous on some targets. */
-//#define __LIBPHOENIX_ARCH_TLS_SUPPORTED
+#define __LIBPHOENIX_ARCH_TLS_SUPPORTED
 
 #endif
