+++ include/arch/armv7a/limits.h
@@ -3,7 +3,7 @@
  *
  * libphoenix
  *
- * Architecture dependent part of limits (arch/armv7)
+ * Architecture dependent part of limits (arch/armv7a)
  *
  * Copyright 2017-2019 Phoenix Systems
  * Author: Pawel Pisarczyk, Aleksander Kaminski, Andrzej Glowinski, Marek Bialowas
@@ -13,8 +13,8 @@
  * %LICENSE%
  */
 
-#ifndef _LIBPHOENIX_ARCH_ARMV7_LIMITS_H_
-#define _LIBPHOENIX_ARCH_ARMV7_LIMITS_H_
+#ifndef _LIBPHOENIX_ARCH_ARMV7A_LIMITS_H_
+#define _LIBPHOENIX_ARCH_ARMV7A_LIMITS_H_
 
 #include "arch.h"
 
@@ -56,10 +56,10 @@
 
 /*** POSIX-required defines ***/
 
-#define PATH_MAX    256  /* Maximum number of bytes the implementation will store as a pathname in a user-supplied buffer of unspecified size, including the terminating null character. MIN: 256 */
-#define NAME_MAX    64   /* Maximum number of bytes in a filename (not including the terminating null of a filename string). MIN: 14 */
-#define ARG_MAX     1500 /* Maximum length of argument to the exec functions including environment data. MIN: 4096 */
-#define SYMLOOP_MAX 8    /* Maximum number of symbolic links that can be reliably traversed in the resolution of a pathname in the absence of a loop. MIN: 8 */
+#define PATH_MAX    1024  /* Maximum number of bytes the implementation will store as a pathname in a user-supplied buffer of unspecified size, including the terminating null character. MIN: 256 */
+#define NAME_MAX    256   /* Maximum number of bytes in a filename (not including the terminating null of a filename string). MIN: 14 */
+#define ARG_MAX     32000 /* Maximum length of argument to the exec functions including environment data. MIN: 4096 */
+#define SYMLOOP_MAX 8     /* Maximum number of symbolic links that can be reliably traversed in the resolution of a pathname in the absence of a loop. MIN: 8 */
 
 #define _POSIX2_RE_DUP_MAX 255
